refactor: function to process message
